requirements.txt
Flask>=2.0
SQLAlchemy>=1.4
alembic>=1.11
Flask==2.3.2
Flask-RESTX==1.3.2
Werkzeug<3.0
Jinja2==3.1.2
pandas==2.2.3
openpyxl==3.1.2
pytest==7.4.0
pytest-asyncio==0.22.0
pytest-xdist==3.7.0
pytest-rerunfailures==12.0
pytest-repeat==0.9.4
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.18
alembic==1.12.0
psycopg2-binary==2.9.10
redis==4.6.0
pydantic==1.10.9
black==24.3.0
flake8==6.0.0
mypy==1.4.1
pre-commit==3.4.0
rapidfuzz>=2.0.0

# Authentication dependencies
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4

# API Security & Validation
Flask-Limiter>=3.13
Flask-JWT-Extended>=4.7.1
marshmallow>=4.0.1
